Finxo
# Finexo Website - Part 1 (HTML & CSS)

This is the source code for the "Finexo Website Part 1" tutorial video on my YouTube channel [Zabiullah Technical](https://www.youtube.com/@ZabiullahTechnical). In this part, we begin creating a modern, responsive website using only HTML and CSS.

## 🔥 What’s Included in Part 1:
- Project folder structure
- HTML layout and basic elements
- CSS styling for header and hero section
- Responsive design setup

## 📺 Watch the Tutorial:
👉 [Watch Part 1 on YouTube](https://www.youtube.com/@ZabiullahTechnical)

## 📁 How to Use:
1. Download or clone this repo
2. Open the folder in Visual Studio Code
3. Run with Live Server

## 📌 Tools Used:
- Visual Studio Code
- Google Fonts
- Font Awesome Icons

---

Stay tuned for Part 2 and beyond. Don't forget to ⭐ star the repo and follow me for more!